# docs/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Betsi Patient Flow API",
    "description": "Patient flow and escalation for emergency departments. Errors are RFC 9457 problem details with a stable \u0027code\u0027. See docs/API.md and docs/API-VERSIONING.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/integrations/sources": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInboundSourceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInboundSourceRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInboundSourceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundSourceSecretIssued"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Integrations"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InboundSourceView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/sources/{sourceId}/rotate-secret": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundSourceSecretIssued"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/sources/{sourceId}/deactivate": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/integrations/sources/{sourceId}/messages": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "summary": "Recent messages from a source, without bodies. Filter by \u0060Quarantined\u0060 to find messages needing review.",
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InboundMessageSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/messages/{messageId}": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "summary": "One message, including the body of a quarantined message. Contains patient data; the read is audited.",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/integrations/inbound/{tenantId}/{sourceId}": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Receives one signed HL7 v2 or FHIR R4 message. Authenticated by the \u0060Betsi-Signature\u0060\nheader, not a bearer token. Requires \u0060Betsi-Message-Id\u0060; resending the same id is safe.",
        "description": "A message that cannot be processed is quarantined and still acknowledged as received (HL7\n\u0060AE\u0060, or \u0060status: Quarantined\u0060), so the sender does not retry a message that will\nnever succeed. Quarantined messages are listed for review.",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundOutcome"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks/register": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Subscribes a URL to event types. The signing secret is in the response and is never shown again.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWebhookRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWebhookRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWebhookRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSecretIssued"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks/{subscriptionId}/rotate-secret": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSecretIssued"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{subscriptionId}/deactivate": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InboundMessageSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "messageId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "messageType": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "InboundOutcome": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "messageType": {
            "type": "string",
            "nullable": true
          },
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "duplicate": {
            "type": "boolean"
          },
          "controlId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InboundSourceSecretIssued": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "format": {
            "type": "string",
            "nullable": true
          },
          "inboundPath": {
            "type": "string",
            "nullable": true
          },
          "secret": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InboundSourceView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "format": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "inboundPath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterInboundSourceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "format": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterWebhookRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WebhookSecretIssued": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "secret": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Returned once, at creation or rotation. The secret is not retrievable afterwards."
      },
...
    },
    "securitySchemes": {
      "X-Betsi-Tenant": {
        "type": "apiKey",
        "description": "Tenant ID. Development tenants: 11111111-1111-1111-1111-111111111111 or 22222222-2222-2222-2222-222222222222.",
        "name": "X-Betsi-Tenant",
        "in": "header"
      },
      "X-Betsi-Actor-Role": {
        "type": "apiKey",
        "description": "Actor role, e.g. Nurse.",
        "name": "X-Betsi-Actor-Role",
        "in": "header"
      },
      "X-Betsi-Actor": {
        "type": "apiKey",
        "description": "Optional actor ID (GUID).",
        "name": "X-Betsi-Actor",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "An access token from the configured OIDC provider. In Development the X-Betsi headers may be used instead.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [],
      "X-Betsi-Tenant": [],
      "X-Betsi-Actor-Role": [],
      "X-Betsi-Actor": []
    }
  ],
...
}